Moved graphing back to test.

+++ src/main/java/knn/api/KnowledgeNode.java
@@ -22,7 +22,7 @@ public class KnowledgeNode {
     private static final double[] SIGMOID_VALUES = {0, 2, 5, 11, 27, 50, 73, 88, 95, 98, 100}; //sigmoid function activation value
     private final Tag inputTag;
     private final Set<Tag> outputTags;  // Integer is the value of confidence
-    private final double threshold; // limit: When activation > threshold : fires output tags (outputFacts array). These tags can be lists of rules or facts.
+    private final int threshold; // limit: When activation > threshold : fires output tags (outputFacts array). These tags can be lists of rules or facts.
     private final int strength; // Which strength approach to take?
     private final double maxAge;
